Home

JuliennedArrays.jl

JuliennedArrays.Reduce

JuliennedArrays.align

JuliennedArrays.julienne

JuliennedArrays.Reduce — Type.

struct Reduce{F}

Reduction of another function. Enables optimizations in some cases.

julia> using JuliennedArrays

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> map(Reduce(+), julienne(array, (*, :)))
3×1 Array{Int64,2}:
 15
  6
 24

julia> array = reshape(1:8, 2, 2, 2)
2×2×2 reshape(::UnitRange{Int64}, 2, 2, 2) with eltype Int64:
[:, :, 1] =
 1  3
 2  4

[:, :, 2] =
 5  7
 6  8

julia> map(Reduce(+), julienne(array, (:, *, :)))
1×2×1 Array{Int64,3}:
[:, :, 1] =
 14  22

source

JuliennedArrays.align — Method.

align(slices, code)

Align an array of slices into a larger array. Code should be a tuple with an entry for each dimension of the desired output. Slices will slide into dimensions coded by :, while * indicates dimensions taken up by the container array. Each slice should be EXACTLY the same size.

julia> using JuliennedArrays, MappedArrays

julia> code = (*, :);

julia> array = [5 6 4; 1 3 2]
2×3 Array{Int64,2}:
 5  6  4
 1  3  2

julia> views = julienne(array, code);

julia> align(mappedarray(sort, views), code)
2×3 Array{Int64,2}:
 4  5  6
 1  2  3

source

JuliennedArrays.julienne — Method.

julienne(array, code)

Slice an array and create views. The code should a tuple of length ndims(array), where : indicates an axis parallel to slices and * axes an axis perpendicular to slices.

julia> using JuliennedArrays

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> map(sum, julienne(array, (*, :)))
3-element Array{Int64,1}:
 15
  6
 24

source