Home

JuliennedArrays.jl

Demo

This package contains a fast iterator for slicing (julienning) arrays.

Index

JuliennedArrays.Reduction

JuliennedArrays.align

JuliennedArrays.julienne

JuliennedArrays.Reduction — Type.

struct Reduction{F}

A reduction of another function. Enables optimizations in some cases.

julia> using JuliennedArrays, Base.Test, Base.Test

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> @inferred map(Reduction(+), julienne(array, (*, :)))
3×1 Array{Int64,2}:
 15
  6
 24

source

JuliennedArrays.align — Method.

align(slices, code)

Align an array of slices into a larger array. Code should be a tuple with an entry for each dimension of the desired output. Slices will slide into dimensions coded by :, while * indicates dimensions taken up by the container array. Each slice should be EXACTLY the same size.

julia> using JuliennedArrays, MappedArrays, Base.Test

julia> code = (*, :);

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> views = @inferred julienne(array, code);

julia> @inferred align(mappedarray(sort, views), code)
3×3 Array{Int64,2}:
 4  5  6
 1  2  3
 7  8  9

source

JuliennedArrays.julienne — Method.

julienne([T = Views], array, code)

Slice an array and create Parts of type T. T should be one of Arrays, Swaps, or Views. It defaults to Views. See the Parts package for more information. The code should a tuple of length ndims(array), where : indicates an axis parallel to slices and * indices an axis perpendicular to slices.

julia> using JuliennedArrays, Base.Test

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> @inferred map(sum, julienne(array, (*, :)))
3-element Array{Int64,1}:
 15
  6
 24

source