Home

JuliennedArrays.jl

Demo

This package contains a fast iterator for slicing (julienning) arrays.

Index

JuliennedArrays.Reduction

JuliennedArrays.align

JuliennedArrays.julienne

JuliennedArrays.Reduction — Type.

struct Reduction{F} <: FunctionOptimization; f::F; end

A reduction of another function. Enables optimizations in some cases. Automatically enabled for known reduction functions.

julia> using JuliennedArrays, Base.Test, Base.Test

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> swaps = @inferred julienne(Swaps, array, (*, :));

julia> @inferred map(Reduction(+), swaps)
3×1 Array{Int64,2}:
 15
  6
 24

julia> @inferred map(sum, swaps)
3×1 Array{Int64,2}:
 15
  6
 24

source

JuliennedArrays.align — Method.

align(slices, code)

Align an array of slices into a larger array. Code should be a tuple for each dimension of the desired output. Slices will slide into dimensions coded by :, while * indicates dimensions taken up by the container array. Each slice should be EXACTLY the same size.

julia> using JuliennedArrays, MappedArrays, Base.Test

julia> code = (*, :);

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> swaps = @inferred julienne(Swaps, array, code);

julia> @inferred align(mappedarray(sort, swaps), code)
3×3 Array{Int64,2}:
 4  5  6
 1  2  3
 7  8  9

source

JuliennedArrays.julienne — Method.

julienne(T, array, code)
julienne(T, array, code, swap)

Slice an array and create shares of type T. T should be one of Arrays, Swaps, or Views. The code should a tuple of length ndims(array), where : indicates an axis parallel to slices and * indices an axis perpendicular to slices.

julia> using JuliennedArrays, Base.Test

julia> code = (*, :);

julia> array = [5 6 4; 1 3 2; 7 9 8]
3×3 Array{Int64,2}:
 5  6  4
 1  3  2
 7  9  8

julia> arrays = @inferred julienne(Arrays, array, (*, :));

julia> @inferred map(mean, arrays)
3-element Array{Float64,1}:
 5.0
 2.0
 8.0

julia> views = @inferred julienne(Views, array, (*, :));

julia> @inferred map(mean, views)
3-element Array{Float64,1}:
 5.0
 2.0
 8.0

julia> swaps = @inferred julienne(Swaps, array, (*, :));

julia> @inferred map(mean, swaps)
3-element Array{Float64,1}:
 5.0
 2.0
 8.0

source