Home

JuliennedArrays.jl

Demo

This package contains a fast iterator for slicing (julienning) arrays.

Index

JuliennedArrays.OutOfPlaceArray

JuliennedArrays.Reduction

JuliennedArrays.Swap

JuliennedArrays.View

Base.map

JuliennedArrays.combine

JuliennedArrays.julienne

JuliennedArrays.OutOfPlaceArray — Type.

struct OutOfPlaceArray{F} <: FunctionOptimization; f::F; end

A function with a faster in-place analog on a whole preallocated output.

source

JuliennedArrays.Reduction — Type.

struct Reduction{F} <: FunctionOptimization; f::F; end

A reduction of another function

source

JuliennedArrays.Swap — Type.

struct Swap{F} <: FunctionOptimization; f::F; end

A function that can be reused on a temporary array

source

JuliennedArrays.View — Type.

struct View{F} <: FunctionOptimization; f::F; end

A function that should be called on views, not copies

source

JuliennedArrays.combine — Method.

combine(pieces)

Combine many pieces of an array.

julia> using JuliennedArrays

julia> array = [1 3 2; 5 6 4; 7 9 8]
3×3 Array{Int64,2}:
 1  3  2
 5  6  4
 7  9  8

julia> result = combine(Base.Generator(sort, julienne(array, (*, :))))
3×3 Array{Int64,2}:
 1  2  3
 4  5  6
 7  8  9

julia> array == result
false

julia> result2 = combine(Base.Generator(View(sort!), julienne(array, (*, :))))
3×3 Array{Int64,2}:
 1  2  3
 4  5  6
 7  8  9

julia> array == result2
true

source

JuliennedArrays.julienne — Method.

julienne(array, julienne_code)

Create a view of an array which will return slices. The julienne code should a tuple of length ndims(array), where : indicates an axis parallel to slices and * indices an axis perpendicular to slices.

julia> using JuliennedArrays

julia> array = reshape([1 3 2; 5 6 4; 7 9 8], 3, 3)
3×3 Array{Int64,2}:
 1  3  2
 5  6  4
 7  9  8

julia> foreach(sort!, julienne(array, (*, :)));

julia> array
3×3 Array{Int64,2}:
 1  2  3
 4  5  6
 7  8  9

source

Base.map — Method.

Base.map(f, r::ReindexedArray)

julia> using JuliennedArrays

julia> array = reshape(1:9, 3, 3)
3×3 Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}:
 1  4  7
 2  5  8
 3  6  9

julia> map(sum, julienne(array, (:, *)))
1×3 Array{Int64,2}:
 6  15  24

julia> map(median, julienne(array, (:, *)))
1×3 Array{Float64,2}:
 2.0  5.0  8.0

julia> map(mean, julienne(array, (:, *)))
1×3 Array{Float64,2}:
 2.0  5.0  8.0

source